Rebound a few vim keys and fixed up a minor bug.
A minor error was preventing the Postgres capitalization from working
with To() as desired in cases where the last letter of a line was part
of a keyword.

Flipped the binding of 'h' and 'l' to match oldschool keyboards and ditched
the mnemoic. Also changed the binding for selecting a whole word as
navigating with 'w' and 'W' saves effort.