Fixed problem w/ diffs & reading/writing scp files
Reverted a setting I've had in place for a long time which caused vim to
drop out (suspend itself) and show the shell whenever a file was loaded
over scp, and whenever two or more panes were diffed, among other
operations.

Following such actions, one or more calls to `fg` were required to
resume using vim, as in the case of vim's `diffthis`, since it
presumably invokes several system utilities as it prepares the panes to
show changes between files. This was annoying to say the least. Glad I
figured it out.

additional keywords: diffing, diffthis