Hack for https://github.com/vim/vim/issues/1268
Basically, remote files loaded via netrw over SCP are downloaded just
fine, but the filename is stored afterwards such that netrw's current
code fails to write the file back to the remote machine when saving IF
the target filename on the remote machine contains spaces (it's not
escaped quite right so SCP complains).

If this gets fixed for netrw, then the kludge written in this commit can
be wiped out or only conditionally applied depending on the installed
version of netrw. Until then, life goes on.