Improvements in vimrc.
Altered robustsave tweak in vimrc to handle writing/saving local files
with spaces in their filenames as this would fail before.

Overrode a new default that manifested in the latest cygwin-vim build
that sets a line-buffer around the cursor so it cannot appear at the
very top or very bottom of the screen. I'm used to the old behavior so I
changed this setting back explicitly.

Swapped away from having the mouse work fully in all the main modes
after having tried it. I seemed to use it in vim much more under Ubuntu
for whatever reason, perhaps because the main keyboard on the Ubuntu
machine I use is a little awkward, which drove me toward the mouse.