Fixed a vim autocommand that I broke in 812135
The striptrailing autocommand-group (augroup) was not actually stripping
trailing whitespace. I had two mistakes wherein (A) I used a
double-quoted string instead of a single-quoted string (which caused
backslashes to be eaten), and then (B) I was using `normal!` to execute
the `:s` (substitution) command, which meant - I believe - I would've
needed a \<Return> in there for it to actually work.