Cool git alias and git hook stuff.
Tweaked the complex git alias (shell function shadow) to always run the
custom pre-commit hook. Wound up doing this because git will complain if
you're working tree lives on a drive that's mounted with `noexec` or
similar.

Also fixed an issue affecting Ubuntu and possibly other systems with the
where the hook-script's call to tput would cause "256" to be printed to
the console, which wasn't harmful, but wasn't helpful either and looked
kinda glitchy.